Each article is now saved to the disk when requested
Ideally an article should only be saved to disk when an editor saves a page with a property or grid control using that particular article. This would however require a lot of extra code and work, so it's easier just to save the article when requested from the Borger.dk web service.